+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+## 1.9.12
+
+- `DBObjectGCSAdapter`:
+  - Added properties: `cacheFilesLimit`, `cacheCheckMaxSkips` and `cacheCheckTimeout`.
+  - `_checkCacheDirectoryLimit`:
+    - Optimize and also use `cacheFilesLimit`.
+    - Log check time.
+
+- reflection_factory: ^2.7.2
+- postgres: ^3.5.9
+- crypto: ^3.0.7
+- http: ^1.6.0
+- _discoveryapis_commons: ^1.0.7
+
+- build_runner: ^2.10.1
+
 ## 1.9.11
 
 - `EntityHandler`: